# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Home Logbook API
  version: 0.11.0
servers:
- url: http://192.168.68.62:8787/api/v1
security:
- bearerAuth: []
paths:
...
  /event-types/{key}/series:
    parameters:
    - in: path
      name: key
      required: true
      schema:
        type: string
    get:
      summary: Return converted values, durations, and per-event hover details
      parameters:
      - in: query
        name: from
        schema:
          type: string
          format: date-time
      - in: query
        name: to
        schema:
          type: string
          format: date-time
      - in: query
        name: bucket
        schema:
          type: string
          default: 1 hour
      - in: query
        name: displayUnitKey
        schema:
          type: string
        description: Exact compatible unit key. Defaults to the event type default.
      - in: query
        name: timeZone
        description: IANA timezone used to align day and week buckets to local calendar boundaries.
        schema:
          type: string
          default: UTC
          examples:
          - America/New_York
      responses:
        '200':
          description: Bucket aggregates plus raw event points with value
          text: null
          note: null
          and duration details: null
        '400':
          description: Incompatible display unit
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...